Extract private classes from TranslationsActivity. Move all related classes into their own package
