+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".activity.TranslationsActivity"
+            android:name=".activity.translations.TranslationsActivity"
             android:screenOrientation="portrait">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
