+++ app/src/main/java/org/mercycorps/translationcards/activity/addTranslation/EnterSourcePhraseActivity.java
@@ -6,7 +6,7 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 
 import org.mercycorps.translationcards.R;
-import org.mercycorps.translationcards.activity.TranslationsActivity;
+import org.mercycorps.translationcards.activity.translations.TranslationsActivity;
 
 import butterknife.Bind;
 import butterknife.OnClick;
