+++ app/src/main/res/layout/activity_translations.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.TranslationsActivity"
+    tools:context=".activity.translations.TranslationsActivity"
     android:background="@color/listBackgroundColor"
     android:id="@+id/mainActivity">
